♻️ Move function to resolve deps to dependency.rs
